movie_manager/app/pyproject.toml
[tool.poetry]
name = "movie-env"
version = "0.1.0"
description = ""
authors = ["igor udot <horw7001@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.5.2"
django = "4.1.3"
django-split-settings = "1.2.0"
isort = "5.10.1"
psycopg2-binary = "2.9.5"
python-dotenv = "0.21.0"
sqlparse = "0.4.3"
django-debug-toolbar = "^3.8.1"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
